🎨 Move type mods next to the code that they represent
